pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gdoc-fetch"
version = "0.1.0"
description = "gdoc-fetch & gdoc-upload: fetch Google Docs to Markdown, upload Markdown to Google Docs (with images)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "markdown>=3.4.0",
    "markdownify>=0.11.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["gdoc_*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
]

[project.scripts]
gdoc-fetch = "gdoc_fetch.cli:main"
gdoc-upload = "gdoc_upload.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]
# Run unit tests by default, skip integration unless explicitly requested
addopts = "-v -m 'not integration'"
python_files = "test_*.py"
python_functions = "test_*"
